+++ src/pages/DataAccessRequest/DataAccessRequest.js
@@ -47,6 +47,7 @@ import ConfirmSubmissionModal from './components/ConfirmSubmissionModal/ConfirmS
 import DeleteDraftModal from './components/DeleteDraftModal/DeleteDraftModal';
 import DuplicateApplicationModal from './components/DuplicateApplicationModal/DuplicateApplicationModal';
 import SelectDatasetModal from './components/SelectDatasetModal/SelectDatasetModal';
+import createCSV from './utils';
 
 import { CSVLink } from 'react-csv';
 
@@ -1584,37 +1585,6 @@ class DataAccessRequest extends Component {
 		}
 	};
 
-	createCSV = (questionAnswers) => {
-		// set headers
-		const headers = ["ID", "Status", "Status notes", "Date submitted", "Last activity", "Datasets", "Project name", "Project submitted by", "Applicant full name", "Job title", "Telephone", "Institutional afflilation", "ORCID", "Email", "Role", "Access to data", "Organisation name", "Line address 1",	"Line address 2", "Postcode", "City", "Country"];
-		const row = [
-			"state._id",
-			"state.applicationStatus",
-			"Bring from screenData",
-			"Bring from screenData",
-			"Bring from screenData",
-			"Bring from screenData",
-			questionAnswers.applicationthisapplication,
-			questionAnswers.safeprojectapplicantfullname,
-			questionAnswers.safeprojectapplicantfullname,
-			questionAnswers.safeprojectapplicantjobtitle,
-			questionAnswers.safeprojectapplicanttelephone,
-			questionAnswers.safeprojectapplicantaffiliation,
-			questionAnswers.safeprojectapplicantorcid,
-			questionAnswers.safeprojectapplicantemail,
-			questionAnswers.safeprojectapplicantrole,
-			questionAnswers.safeprojectapplicantdataaccess,
-			questionAnswers.safeprojectorganisationname,
-			questionAnswers.safeprojectorganisationaddress1,
-			questionAnswers.safeprojectorganisationaddress2,
-			questionAnswers.safeprojectorganisationpostcode,
-			questionAnswers.safeprojectorganisationcity,
-			questionAnswers.safeprojectorganisationcountry,
-		];
-
-		return [headers, row];
-	}
-
 	render() {
 		const {
 			lastSaved,
@@ -1979,7 +1949,7 @@ class DataAccessRequest extends Component {
 					appToCloneId={this.state._id}
 			/>
 
-				<CSVLink data={this.createCSV(this.state.questionAnswers)} filename={"test"}>
+				<CSVLink data={createCSV(this.state.questionAnswers)} filename={"test"}>
 					<button className={`button-secondary`}>
 						Download Request as a CSV
 					</button>
